Removing console output from ehvi + Slight code refactoring in experimental + Moved acquiopt tag from bo_multi.hpp since it seems to be algorithm specific
